fix model_helper.py

+++ src/utils/model_helper.py
@@ -129,8 +129,8 @@ def create_or_load_embed(embed_name, vocab_file, embed_file, vocab_size, embed_s
         else:
             embedding, embed_size = _create_pretrained_emb_from_txt_const(vocab_file, embed_file)
     else:
-        unk_embed = tf.Variable(np.random.uniform(-0.1, 0.1, size=[1, embed_size]), dtype=dtype, trainable=True)
-        pad_embed = tf.Variable(np.zeros(shape=(1, embed_size), dtype=dtype), trainable=False)
+        unk_embed = tf.Variable(np.random.uniform(-0.1, 0.1, size=[1, embed_size]), dtype=np.float32, trainable=True)
+        pad_embed = tf.Variable(np.zeros(shape=(1, embed_size)), dtype=dtype, trainable=False)
         other_embedding = tf.get_variable(
             embed_name, [vocab_size-2, embed_size], dtype,
             initializer=tf.random_uniform_initializer(-0.1, 0.1), trainable=True)
